+++ eRestoran/eRestoran.WebAPI/Database/Jelo.cs
@@ -10,7 +10,7 @@ namespace eRestoran.WebAPI.Database
 		public int Id { get; set; }
 		public string Naziv { get; set; }
 		public string Opis { get; set; }
-		public int Cijena { get; set; }
+		public decimal Cijena { get; set; }
 
 		public int KategorijaId { get; set; }
 		public Kategorija Kategorija { get; set; }
